refactor: todoStore.subscribe 함수 위치를 ColumnBody로 변경
